feat: create new version of each resource in multi-resource deployment
This commit changes the deployment logic if multiple resources are deployed together:
If at least one of them has changed (compared to the latest already deployed version), a new version of _every_ resource
contained in the deployment will be created (even if this particular resource has not changed).

The change is done to properly support the new "deployment" binding type:
If a process contains a resource link with "deployment" binding, each instance of a particular version of this process
should call the same version of the target resource (to ensure a deterministic behavior that is easy to understand).

For additional context see https://github.com/camunda/camunda/issues/19859.